requirements.txt
Flask>=2.0
pymongo>=4.0
python-dotenv
pillow
razorpay
gridfs
boto3
requests
